Updated docs to add yarn install note
This commit updates the documentation to cover a bug in how yarn install notes explain how to set your path and how globally installed binaries are added to a system. Hopefully in the future the getting started notes will explain that if you install binaries globally on a system you need to add that global path to your environment path.